requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.0
undetected-chromedriver>=3.5.0
rich>=13.0.0
tqdm>=4.65.0
Pillow>=10.0.0
requests>=2.31.0
pandas>=2.0.0
